Lock and Key System
The cube that could be used to pick up items must now be used to open the door. Just pick it up and push it into the first door. You still need the code as well.
I have only done the first door. The second door can be done later.